Move Apprentice repository to IContext
