+++ src/SFA.DAS.ApprenticeCommitments/Application/Commands/ChangeEmailAddressCommand/ChangeEmailAddressCommandHandler.cs
@@ -8,9 +8,9 @@ namespace SFA.DAS.ApprenticeCommitments.Application.Commands.CreateRegistrationC
 {
     public class ChangeEmailAddressCommandHandler : IRequestHandler<ChangeEmailAddressCommand>
     {
-        private readonly ApprenticeRepository _apprenticeRepository;
+        private readonly IApprenticeContext _apprenticeRepository;
 
-        public ChangeEmailAddressCommandHandler(ApprenticeRepository apprenticeRepository)
+        public ChangeEmailAddressCommandHandler(IApprenticeContext apprenticeRepository)
             => _apprenticeRepository = apprenticeRepository;
 
         public async Task<Unit> Handle(ChangeEmailAddressCommand command, CancellationToken cancellationToken)
