+++ src/SFA.DAS.ApprenticeCommitments/Data/IApprenticeContext.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.EntityFrameworkCore;
+using SFA.DAS.ApprenticeCommitments.Data.Models;
+using SFA.DAS.ApprenticeCommitments.Exceptions;
+using System;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.ApprenticeCommitments.Data
+{
+    public interface IApprenticeContext : IEntityContext<Apprentice>
+    {
+        internal async Task<Apprentice> GetById(Guid apprenticeId)
+            => await Find(apprenticeId)
+                    ?? throw new DomainException(
+                        $"Apprentice {apprenticeId} not found");
+
+        private async Task<Apprentice> Find(Guid apprenticeId)
+            => await Entities.SingleOrDefaultAsync(a => a.Id == apprenticeId);
+    }
+}
