+++ src/SFA.DAS.ApprenticeCommitments/Data/Models/ApprenticeCommitmentsDbContext.cs
@@ -4,7 +4,7 @@ using System.Net.Mail;
 
 namespace SFA.DAS.ApprenticeCommitments.Data.Models
 {
-    public class ApprenticeCommitmentsDbContext : DbContext, IRegistrationContext, IApprenticeshipContext
+    public class ApprenticeCommitmentsDbContext : DbContext, IRegistrationContext, IApprenticeContext, IApprenticeshipContext
     {
         public ApprenticeCommitmentsDbContext()
         {
@@ -19,6 +19,7 @@ namespace SFA.DAS.ApprenticeCommitments.Data.Models
         public virtual DbSet<Apprenticeship> Apprenticeships { get; set; }
 
         DbSet<Registration> IEntityContext<Registration>.Entities => Registrations;
+        DbSet<Apprentice> IEntityContext<Apprentice>.Entities => Apprentices;
         DbSet<Apprenticeship> IEntityContext<Apprenticeship>.Entities => Apprenticeships;
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
