+++ src/SFA.DAS.ApprenticeCommitments/Infrastructure/ServiceCollectionExtensions.cs
@@ -40,8 +40,8 @@ namespace SFA.DAS.ApprenticeCommitments.Infrastructure
             services.AddSingleton<IManagedIdentityTokenProvider, ManagedIdentityTokenProvider>();
             services.AddTransient<IConnectionFactory, SqlServerConnectionFactory>();
             services.AddScoped<IRegistrationContext>(s => s.GetRequiredService<ApprenticeCommitmentsDbContext>());
+            services.AddScoped<IApprenticeContext>(s => s.GetRequiredService<ApprenticeCommitmentsDbContext>());
             services.AddScoped<IApprenticeshipContext>(s => s.GetRequiredService<ApprenticeCommitmentsDbContext>());
-            services.AddScoped<ApprenticeRepository>();
 
             return services;
         }
